Update Dockerfile 2025-03-01
gcc: The GNU Compiler Collection is necessary for compiling C programs. It’s commonly used alongside g++ to handle C and C++ compilation, respectively.

make: This is a build automation tool used in conjunction with gcc and g++ to automate the compilation process via Makefiles. It is an essential tool when working on C/C++ projects.

python3-dev: This package includes the necessary header files and libraries required to build Python C extensions. It's crucial if you're installing or compiling Python modules that require C extensions or interfacing with C/C++ code (such as numpy, pandas, etc.).